+++ src/hooks/hooks.tsx
@@ -2,6 +2,8 @@ import { useState, useEffect, useRef } from 'react';
 import { isEqual, omit } from 'lodash';
 import { UserProfile } from '@lib/classes';
 import { getProfilePrefix, validateEmail, validatePassword, validateProfileSlug } from '@lib/utils';
+import { As, AvatarBadge, Icon } from '@chakra-ui/react';
+import { FiCalendar } from 'react-icons/fi';
 
 /**
  * Custom hooks.
@@ -105,6 +107,10 @@ export const useErrorMessage = (errorCode?: string, defaultMessage: string = 'Er
 		case 'user_slug_invalid':
 			return 'Only letters, numbers, dashes (-) and underscores (_) are allowed.';
 
+		// Other errors
+		case 'conflict_range_overlap':
+			return 'This date range overlaps with an existing busy time. Please try again.';
+
 		default:
 			return defaultMessage + ': ' + errorCode;
 	}
@@ -120,8 +126,9 @@ export const useProfileEdited = (editProfile: UserProfile, origProfile: UserProf
 	if (origProfile === null) return;
 
 	const ignoreFields = [
-		'slug',
 		'credits',
+		'conflictRanges',
+		'slug',
 		'image',
 		'resume',
 		'mediaImage1',