+++ src/lib/types.ts
@@ -1,4 +1,4 @@
-import { WPItem } from '@lib/classes';
+import { DateRange, WPItem } from '@lib/classes';
 
 /**
  * The data shape for generic WordPress item input.
@@ -115,6 +115,15 @@ export interface CreditParams {
 }
 
 /**
+ * The data shape for DateRange input.
+ */
+export interface DateRangeParams {
+	id?: number;
+	startDate?: string | Date | undefined;
+	endDate?: string | Date | undefined;
+}
+
+/**
  * The data shape for Credit output to GraphQL.
  */
 export interface CreditOutput {
@@ -188,6 +197,7 @@ export interface SearchFilterSet {
 		jobs?: string[];
 	};
 	skills?: string[];
+	jobDates?: DateRange;
 	unions?: string[];
 	locations?: string[];
 	experienceLevels?: string[];
@@ -195,10 +205,12 @@ export interface SearchFilterSet {
 	racialIdentities?: string[];
 	personalIdentities?: string[];
 	searchName?: string;
+	searchDates?: DateRangeParams;
 }
 
 /**
- * The data shape for a raw search query (flat `positions`).
+ * The data shape for a raw search query.
+ * Flattens `positions` and omits `jobDates`.
  */
 export interface SearchFilterSetRaw {
 	positions?: string[];
@@ -221,6 +233,7 @@ export interface SearchFilterSet {
 		jobs?: string[];
 	};
 	skills?: string[];
+	jobDates?: DateRange;
 	unions?: string[];
 	locations?: string[];
 	experienceLevels?: string[];