+++ src/lib/utils.ts
@@ -3,8 +3,13 @@
  */
 
 import { isEqual } from 'lodash';
-import { Credit, PersonalLinks, UserProfile, WPItem } from '@lib/classes';
-import { SearchFilterSet, SearchFilterSetRaw, SearchResultCandidate } from '@lib/types';
+import { Credit, PersonalLinks, DateRange, UserProfile, WPItem } from '@lib/classes';
+import {
+	DateRangeParams,
+	SearchFilterSet,
+	SearchFilterSetRaw,
+	SearchResultCandidate,
+} from '@lib/types';
 import Cookies from 'js-cookie';
 import { passwordStrength } from 'check-password-strength';
 const { VITE_FRONTEND_URL } = import.meta.env;
@@ -104,7 +109,7 @@ export function getWPItemsFromIds(ids: number[], items: WPItem[]): WPItem[] {
  * @param {UserProfile} profile The user profile to preprae.
  * @returns {Object} The prepared user profile.
  */
-export const prepareUserProfileForGraphQL = (profile: UserProfile): object => {
+export function prepareUserProfileForGraphQL(profile: UserProfile): object {
 	// Strip unwanted fields from the payload
 	const {
 		slug,
@@ -116,12 +121,13 @@ export const prepareUserProfileForGraphQL = (profile: UserProfile): object => {
 		mediaImage4,
 		mediaImage5,
 		mediaImage6,
+		conflictRanges,
 		credits,
 		...sanitized
 	} = profile;
 
 	return sanitized;
-};
+}
 
 /**
  * Sort two arrays and compare them.
@@ -221,13 +227,13 @@ export const sortWPItemsByName = (a: WPItem, b: WPItem): number => {
  * @param item The item to toggle.
  * @returns The new array.
  */
-export const toggleArrayItem = (array: any[], item: any): any[] => {
+export function toggleArrayItem(array: any[], item: any): any[] {
 	if (array.includes(item)) {
 		return array.filter((i) => i !== item);
 	}
 
 	return [...array, item];
-};
+}
 
 /**
  * Extract the IDs of terms from a JSON string.
@@ -312,16 +318,19 @@ export function prepareSearchFilterSet(searchObj: any, terms: WPItem[]): SearchF
 
 /**
  * Flattens the `positions` object to a single `positions` property
- * containing an array of job IDs.
+ * containing an array of job IDs, and removes the `jobDates` property.
  *
  * @param searchObj
  * @returns The flattened search object.
  */
-export function flattenfilterSetPositions(searchObj: SearchFilterSet): SearchFilterSetRaw {
+export function prepareSearchFilterSetForSave(searchObj: SearchFilterSet): SearchFilterSetRaw {
 	const preparedSearchObj: SearchFilterSetRaw = {
 		...searchObj,
 		positions: searchObj.positions.jobs,
 	};
+
+	// Remove the `jobDates` property if it exists
+	delete preparedSearchObj.jobDates;
 
 	return preparedSearchObj;
 }
@@ -374,7 +383,7 @@ export function deleteCookie(name: string): void {
 }
 
 /**
- * Prepare a collection of Credit objects from as raw array of GQL nodes.
+ * Prepare a collection of Credit objects from a raw array of GQL nodes.
  *
  * @param nodes The raw array of GQL nodes to prepare.
  * @returns An array of Credit objects.
@@ -425,6 +434,16 @@ export function prepareCreditsFromGQLNodes(nodes: object[]): Credit[] {
 	});
 
 	return credits;
+}
+
+/**
+ * Generates an array of DateRange objects from the provided array of GraphQL nodes.
+ *
+ * @param {object[]} nodes - An array of GraphQL nodes to process.
+ * @return {DateRange[]} An array of DateRange objects prepared from the GraphQL nodes.
+ */
+export function prepareUnavailDatesFromGQLNodes(nodes: DateRangeParams[]): DateRange[] {
+	return nodes.map((node: DateRangeParams) => new DateRange(node));
 }
 
 /**
@@ -501,4 +520,25 @@ export function obscureEmail(emailString: string): string {
 	const obscuredPart = '*'.repeat(username.length - visibleCount);
 
 	return `${username.substring(0, visibleCount)}${obscuredPart}@${domain}`;
+}
+
+/**
+ * Check if a job date search filter range overlaps with any of the given conflict ranges.
+ *
+ * @param jobDates - The job schedule to check.
+ * @param conflictRange - An array of conflict ranges to check against.
+ * @returns True if the job schedule overlaps with any of the conflict ranges, false otherwise.
+ */
+export function dateRangesOverlap(jobDates: DateRange, conflictRange: DateRange): boolean {
+	const { startDate: jobStart, endDate: jobEnd } = jobDates || {};
+	const { startDate: rangeStart, endDate: rangeEnd } = conflictRange;
+
+	if (!jobStart || !rangeStart || !rangeEnd) return false;
+
+	// If the job does not have an endDate, we only return true if the conflict range overlaps with jobStart.
+	if (!jobEnd) {
+		return jobStart <= rangeEnd;
+	}
+
+	return rangeStart <= jobEnd && rangeEnd >= jobStart;
 }