+++ changelog.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file, starting
 with 1.0.0beta.
 
+## \[1.1.7-avail-calendar] - Unreleased
+
+- Add: Conflict Dates calendar and job search dates.
+- Improve: Remove redundant `sortCreditsByIndex()` from useUserProfile hook.
+- Improve: Saved Search styling
+- Change: Reusing a saved search no longer automatically searches, but rather populates filters for re-running when ready.
+- Change: "Quick Actions" --> "Options".
+
 ## \[1.1.7] - Unreleased
 
 - Improve: @common component file organization